large refactoring. change directory, modified test, lib class...
